requirements.txt
fastapi
uvicorn
sqlalchemy
aiosqlite
httpx
python-jose[cryptography]
passlib
pytest-mock==3.12.0
python-dotenv==1.0.0
python-multipart
python-dateutil
httpx>=0.23.0
tenacity==8.2.3